# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_regenerate_note_chunk_api_jobs__job_id__note_chunks__chunk_id__regenerate_post": {
        "properties": {
          "extras": {
            "default": "",
            "title": "Extras",
            "type": "string"
          },
          "frame_limit": {
            "default": 6,
            "title": "Frame Limit",
            "type": "integer"
          },
          "note_api_key": {
            "title": "Note Api Key",
            "type": "string"
          },
          "note_base_url": {
            "default": "https://api.openai.com/v1",
            "title": "Note Base Url",
            "type": "string"
          },
          "note_language": {
            "$ref": "#/components/schemas/NoteLanguage",
            "default": "zh"
          },
          "note_model": {
            "default": "gpt-5.5",
            "title": "Note Model",
            "type": "string"
          },
          "note_style": {
            "$ref": "#/components/schemas/NoteStyle",
            "default": "detailed"
          }
        },
        "required": [
          "note_api_key"
        ],
        "title": "Body_regenerate_note_chunk_api_jobs__job_id__note_chunks__chunk_id__regenerate_post",
        "type": "object"
      },
      "Body_regenerate_note_version_endpoint_api_jobs__job_id__note_versions_post": {
        "properties": {
          "extras": {
            "default": "",
            "title": "Extras",
            "type": "string"
          },
          "frame_limit": {
            "default": 6,
            "title": "Frame Limit",
            "type": "integer"
          },
          "note_api_key": {
            "default": "",
            "title": "Note Api Key",
            "type": "string"
          },
          "note_base_url": {
            "default": "https://api.openai.com/v1",
            "title": "Note Base Url",
            "type": "string"
          },
          "note_language": {
            "$ref": "#/components/schemas/NoteLanguage"
          },
          "note_model": {
            "default": "gpt-5.5",
            "title": "Note Model",
            "type": "string"
          },
          "note_style": {
            "$ref": "#/components/schemas/NoteStyle",
            "default": "detailed"
          }
        },
        "required": [
          "note_language"
        ],
        "title": "Body_regenerate_note_version_endpoint_api_jobs__job_id__note_versions_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NoteChunkIndex": {
        "properties": {
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/NoteChunkMeta"
            },
            "title": "Chunks",
            "type": "array"
          },
          "total_segments": {
            "default": 0,
            "title": "Total Segments",
            "type": "integer"
          }
        },
        "title": "NoteChunkIndex",
        "type": "object"
      },
      "NoteChunkMeta": {
        "properties": {
          "end_time": {
            "title": "End Time",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "segment_end": {
            "title": "Segment End",
            "type": "integer"
          },
          "segment_start": {
            "title": "Segment Start",
            "type": "integer"
          },
          "start_time": {
            "title": "Start Time",
            "type": "number"
          },
          "status": {
            "default": "succeeded",
            "title": "Status",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "index",
          "total",
          "label",
          "start_time",
          "end_time",
          "segment_start",
          "segment_end"
        ],
        "title": "NoteChunkMeta",
        "type": "object"
      },
      "NoteLanguage": {
        "enum": [
          "zh",
          "en",
          "follow"
        ],
        "title": "NoteLanguage",
        "type": "string"
      },
      "NoteStyle": {
        "enum": [
          "minimal",
          "detailed",
          "tutorial",
          "academic",
          "task_oriented",
          "meeting_minutes"
        ],
        "title": "NoteStyle",
        "type": "string"
      },
      "NoteVersion": {
        "properties": {
          "active": {
            "default": false,
            "title": "Active",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "draft_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Path"
          },
          "evidence_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Path"
          },
          "extras_length": {
            "default": 0,
            "title": "Extras Length",
            "type": "integer"
          },
          "extras_present": {
            "default": false,
            "title": "Extras Present",
            "type": "boolean"
          },
          "frame_dir": {
            "title": "Frame Dir",
            "type": "string"
          },
          "frame_limit": {
            "title": "Frame Limit",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "note_base_url": {
            "title": "Note Base Url",
            "type": "string"
          },
          "note_language": {
            "title": "Note Language",
            "type": "string"
          },
          "note_model": {
            "title": "Note Model",
            "type": "string"
          },
          "note_path": {
            "title": "Note Path",
            "type": "string"
          },
          "note_style": {
            "$ref": "#/components/schemas/NoteStyle"
          },
          "selected": {
            "default": true,
            "title": "Selected",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "label",
          "note_style",
          "note_language",
          "note_model",
          "note_base_url",
          "frame_limit",
          "note_path",
          "frame_dir"
        ],
        "title": "NoteVersion",
        "type": "object"
      },
      "NoteVersionIndex": {
        "properties": {
          "active_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Version Id"
          },
          "selected_version_ids": {
            "items": {
              "type": "string"
            },
            "title": "Selected Version Ids",
            "type": "array"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/NoteVersion"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "title": "NoteVersionIndex",
        "type": "object"
      },
      "NoteVersionSelection": {
        "properties": {
          "active_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Version Id"
          },
          "selected_version_ids": {
            "items": {
              "type": "string"
            },
            "title": "Selected Version Ids",
            "type": "array"
          }
        },
        "title": "NoteVersionSelection",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Video Note Generator MVP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs/{job_id}/note-chunks": {
      "get": {
        "operationId": "list_note_chunks_api_jobs__job_id__note_chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteChunkIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Note Chunks",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/jobs/{job_id}/note-chunks/{chunk_id}/regenerate": {
      "post": {
        "operationId": "regenerate_note_chunk_api_jobs__job_id__note_chunks__chunk_id__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "chunk_id",
            "required": true,
            "schema": {
              "title": "Chunk Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_regenerate_note_chunk_api_jobs__job_id__note_chunks__chunk_id__regenerate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Regenerate Note Chunk Api Jobs  Job Id  Note Chunks  Chunk Id  Regenerate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Note Chunk",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/jobs/{job_id}/note-versions": {
      "get": {
        "operationId": "list_note_versions_api_jobs__job_id__note_versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteVersionIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Note Versions",
        "tags": [
          "notes"
        ]
      },
      "patch": {
        "operationId": "update_note_version_selection_api_jobs__job_id__note_versions_patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_state_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected State Revision"
            }
          },
          {
            "in": "query",
            "name": "expected_artifact_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Artifact Revision"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteVersionSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteVersionIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Note Version Selection",
        "tags": [
          "notes"
        ]
      },
      "post": {
        "operationId": "regenerate_note_version_endpoint_api_jobs__job_id__note_versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_regenerate_note_version_endpoint_api_jobs__job_id__note_versions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Regenerate Note Version Endpoint Api Jobs  Job Id  Note Versions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Note Version Endpoint",
        "tags": [
          "notes"
        ]
      }
    },
...
    "/api/jobs/{job_id}/preview/note/{version_id}": {
      "get": {
        "operationId": "preview_note_version_api_jobs__job_id__preview_note__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Note Version",
        "tags": [
          "downloads"
        ]
      }
    },
...
  }
}